Move grammar menu to testcreation component and remove blur
