+++ src/components/TestCreation.js
@@ -5,15 +5,29 @@ import Button from './ui/Button'
 import { createTest } from '../lib/firebase'
 import List from './wordpair/List'
 import Form from './wordpair/Form'
+import GrammarMenu from './grammarmenu/GrammarMenu'
+import { ReactComponent as WordClassSVG } from '../icons/wordClass.svg'
 
 export default function TestCreation({ db }) {
   const [wordPairs, setWordPairs] = useState([])
   const [testURL, setTestURL] = useState()
+  const [wordClass, setWordClass] = useState()
+  const [gender, setGender] = useState()
+  const [person, setPerson] = useState()
+  const [number, setNumber] = useState()
+  const [showGrammarMenu, setShowGrammarMenu] = useState(false)
 
   function deleteWord(index) {
     const newWords = [...wordPairs]
     newWords.splice(index, 1)
     setWordPairs(newWords)
+  }
+
+  function changeWordClass(wordClass) {
+    setWordClass(wordClass)
+    setGender()
+    setPerson()
+    setNumber()
   }
 
   return (
@@ -27,6 +41,9 @@ export default function TestCreation({ db }) {
             <List words={wordPairs} deleteWord={deleteWord} />
           )}
           <StickyWrapper>
+            <StyledWordClassSVG
+              onClick={() => setShowGrammarMenu(!showGrammarMenu)}
+            />
             <Form wordPairs={wordPairs} setWordPairs={setWordPairs} />
             <Button
               buttonstyle="primary"
@@ -34,6 +51,19 @@ export default function TestCreation({ db }) {
             >
               Test erstellen
             </Button>
+            {showGrammarMenu && (
+              <GrammarMenu
+                wordClass={wordClass}
+                setWordClass={changeWordClass}
+                gender={gender}
+                setGender={setGender}
+                person={person}
+                setPerson={setPerson}
+                number={number}
+                setNumber={setNumber}
+                setShowGrammarMenu={setShowGrammarMenu}
+              />
+            )}
           </StickyWrapper>
         </StyledTestCreation>
       )}
@@ -54,4 +84,8 @@ const StickyWrapper = styled.div`
   flex-direction: column;
   align-items: center;
   padding-bottom: 10px;
+`
+
+const StyledWordClassSVG = styled(WordClassSVG)`
+  cursor: pointer;
 `
