+++ src/components/wordpair/Form.js
@@ -2,16 +2,9 @@ import React, { createRef, useState } from 'react'
 import styled from 'styled-components'
 import Input from '../ui/Input'
 import Button from '../ui/Button'
-import GrammarMenu from '../grammarmenu/GrammarMenu'
-import { ReactComponent as WordClassSVG } from '../../icons/wordClass.svg'
 
-export default function Form({ wordPairs, setWordPairs }) {
-  const [showGrammarMenu, setShowGrammarMenu] = useState(false)
+export default function Form({ wordPairs, setWordPairs, isBlured }) {
   const [formInput, setFormInput] = useState({ foreign: '', native: '' })
-  const [wordClass, setWordClass] = useState()
-  const [gender, setGender] = useState()
-  const [person, setPerson] = useState()
-  const [number, setNumber] = useState()
 
   const foreignInput = createRef()
 
@@ -31,19 +24,9 @@ export default function Form({ wordPairs, setWordPairs }) {
     setFormInput(newFormInput)
   }
 
-  function changeWordClass(wordClass) {
-    setWordClass(wordClass)
-    setGender()
-    setPerson()
-    setNumber()
-  }
-
   return (
     <>
-      <StyledForm onSubmit={onSubmit} isBlured={showGrammarMenu}>
-        <StyledWordClassSVG
-          onClick={() => setShowGrammarMenu(!showGrammarMenu)}
-        />
+      <StyledForm onSubmit={onSubmit} isBlured={isBlured}>
         <Input
           name="foreign"
           type="text"
@@ -63,19 +46,6 @@ export default function Form({ wordPairs, setWordPairs }) {
           Add word
         </Button>
       </StyledForm>
-      {showGrammarMenu && (
-        <GrammarMenu
-          wordClass={wordClass}
-          setWordClass={changeWordClass}
-          gender={gender}
-          setGender={setGender}
-          person={person}
-          setPerson={setPerson}
-          number={number}
-          setNumber={setNumber}
-          setShowGrammarMenu={setShowGrammarMenu}
-        />
-      )}
     </>
   )
 }
@@ -84,16 +54,6 @@ const StyledForm = styled.form`
   display: flex;
   flex-direction: column;
   align-items: center;
-  ${(props) =>
-    props.isBlured &&
-    `
-      filter: blur(2px); 
-      pointer-events: none;
-    `}
   background: var(--background);
   width: 100%;
-`
-
-const StyledWordClassSVG = styled(WordClassSVG)`
-  cursor: pointer;
 `
